Widen lifetime faith accumulator to 64-bit
m_iFaithEverGeneratedTimes100 is a monotonic lifetime accumulator: it
is only ever increased, guarded in SetFaithTimes100 to add on rising
values alone. As an int it wraps at 21.47M lifetime faith.

Its direct sibling m_lJONSCultureEverGeneratedTimes100 was widened to
long long for exactly this reason and the two are structurally
identical, six lines apart in both the member block and the sync
archive. This restores that consistency.

Consumers are the two minor-civ CONTEST_FAITH quest sites, which store
into an int quest field and so keep an explicit narrowing cast; their
behaviour is unchanged for all values that fit today.

SAVEGAME INCOMPATIBLE: the serialized field changes width, so this must
ship in a Y-level release per NEW RELEASE INSTRUCTIONS.txt.

+++ CvGameCoreDLL_Expansion2/CvMinorCivAI.cpp
@@ -1281,7 +1281,7 @@ int CvMinorCivQuest::GetContestValueForPlayer(PlayerTypes ePlayer) const
 		else
 		{
 			int iStartFaith = pMinor->GetMinorCivAI()->GetQuestData1(ePlayer, eType);
-			int iEndFaith = GET_PLAYER(ePlayer).GetFaithEverGeneratedTimes100();
+			int iEndFaith = (int)GET_PLAYER(ePlayer).GetFaithEverGeneratedTimes100();
 			iValue = iEndFaith - iStartFaith;
 		}
 		break;
@@ -2682,7 +2682,7 @@ void CvMinorCivQuest::DoStartQuest(int iStartTurn, PlayerTypes pCallingPlayer)
 		}
 		else
 		{
-			m_iData1 = pAssignedPlayer->GetFaithEverGeneratedTimes100();
+			m_iData1 = (int)pAssignedPlayer->GetFaithEverGeneratedTimes100();
 		}
 
 		int iTurnsRemaining = GetEndTurn() - GC.getGame().getGameTurn();
